projects/capstone/starter/requirements.txt
alembic==1.3.2
Click==7.0
ecdsa==0.15
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycryptodome==3.3.1
pyflakes==2.1.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==2.0.0
six==1.14.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
